Fix monkeypatching of AdministrativeTags.
